feature/ Moved home section subfolder to components
