+++ src/components/Video.vue
@@ -1,7 +1,7 @@
 <template>
   <main class="relative h-[60vh] sm:h-[80vh]  mt-28">
     <video controls class=" max-w-[343px] max-h-[221px]  sm:max-w-[715px]  sm:max-h-[339px] lg:max-w-[915px] lg:max-h-[439px] mx-auto absolute video">
-      <source src="./video.mp4" type="video/mp4" />
+      <source src="src\assets\video.mp4" type="video/mp4" />
     </video>
     <div class="h-1/2 bg-unify-beige">
