--- src/views/Home/HomeView.vue
@@ -1,10 +1,0 @@
-<template>
-  <LandingVue />
-  <VideoVue />
-  <EveryoneVue />
-</template>
-<script setup>
-import LandingVue from './Sections/Landing.vue';
-import VideoVue from './Sections/Video.vue';
-import EveryoneVue from "./Sections/Everyone.vue"
-</script>
