+++ src/views/HomeView.vue
@@ -1,0 +1,14 @@
+<template>
+  <LandingVue />
+  <VideoVue />
+  <EveryoneVue />
+</template>
+<script setup>
+// import LandingVue from './Home/Sections/Landing.vue';
+// import VideoVue from './Home/Sections/Video.vue';
+// import EveryoneVue from "./Home/Sections/Everyone.vue"
+
+import LandingVue from '../components/Landing.vue';
+import VideoVue from '../components/Video.vue';
+import EveryoneVue from '../components/Everyone.vue';
+</script>
